requirements.txt
asgiref==3.3.4
attrs==21.2.0
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
